[0.0.1+5] Update example project
