+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.0.1+5
+
+* Update example project.
+
 ## 0.0.1+4
 
-* Update readme
+* Update readme.
 
 ## 0.0.1+3
